// src/components/burger-ingredients/lib/types.ts
export interface Ingredient {
  _id: string;
  name: string;
  type: "bun" | "sauce" | "main";
  proteins: number;
  fat: number;
  carbohydrates: number;
  calories: number;
  price: number;
  image: string;
  image_mobile: string;
  image_large: string;
  __v: number;
}

export type IngredientType = "bun" | "sauce" | "main";

export interface TabData {
  value: IngredientType;
  label: string;
  active: boolean;
}